requirements.txt
discord.py
httpx
openai
pyyaml
PyNaCl
yt_dlp
